impl: wire conformance runner via vitest; tolerate typed handshake rejection
Conformance suite now 20/20 green locally against real server + Toxiproxy
(~31s wall clock). Fixes:

1. Runner runs under vitest — server-core/test-utils re-exports
   expectRpcFailure which imports vitest; hosting under vitest sets the
   globals. scripts/test-conformance.ts removed; replaced by
   scripts/conformance.entry.test.ts + vitest.conformance.config.ts.
   test:conformance npm script switched to
   'vitest run --config vitest.conformance.config.ts'. Default
   vitest.config scopes pnpm test to src/**/*.test.ts, so AC12 is
   unchanged (pnpm test stays 61/61).

2. Auto-connect handshake now catches TestingRpcResponseError — fast-
   check's A1 draw of auth/connect with empty strings caused a typed
   server rejection to fail the handshake; tolerating it lets A1 run
   the shape-check it was supposed to.

3. @moltzap/server-core added as protocol devDep for the dynamic
   import. AC13 preserved: the import lives in packages/protocol/
   scripts/, outside src/testing/.

4. .gitignore: conformance-artifacts/ (transient failure-replay dir).

Regression: pnpm test 61/61; tsc --noEmit clean; sloppy-code-guard
all-green.
